update views and move new controllers to v1
